backend/requirements.txt
# FastAPI и зависимости
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3  # для async SQLAlchemy

# Pydantic для валидации
pydantic==2.5.3
pydantic-settings==2.1.0

# JWT и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP клиенты
httpx==0.26.0
aiohttp==3.9.1

# Celery для async задач
celery==5.3.6
redis==5.0.1

# WebSocket
websockets==12.0

# Telegram
python-telegram-bot==20.8

# Обработка изображений
Pillow==10.2.0
python-magic==0.4.27

# Логирование и мониторинг
sentry-sdk==1.40.0

# Тестирование
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
# httpx уже объявлен выше (строка 22)

# Дополнительно
tenacity==8.2.3  # для retry логики
